Fix CAN interface initialization on newer Linux versions
Canny is unable to bind the CAN socket to any CAN interfaces on newer
Linux versions, rendering it unable to initialize.
The problem occurs because the address family in the address struct is
not set before calling `bind()', thus causing EINVAL to be returned on
more recent Linux versions.

This commit modifies the CAN socket initialization so that the address
family is set properly before attempting to bind to an interface.