Fix indentation, formatting, and trailing spaces
Indentation and code formatting of the source code is not uniform,
making it somewhat unpleasant to read. Further, there are a lot of
trailing spaces in license headers and other places that cause line
breaks where none are necessary.

This commit makes indentation and formatting more uniform and
removes trailing spaces.