Add functions for better log messages
Canny does not generate a lot of log output and the output that is
generated does not include timestamps, or log tags to filter output
by. Better logging facilities would make it easier to improve log
outputs in the future.

This commit adds some simple logging functions that can be used to
log messages at different log levels. Log messages are prefixed with
a timestamp and a log tag, allowing messages to be correlated and
filtered. The log level can be configured on the command line using
the `--verbose' and `--quiet' options.