Introduce base type for connections and derive other types from it
Instead of using functions to structure the code in a meaningful way,
canny has a lot going on directly in the main function, which makes
it hard to debug and modify the source.

This commit introduces a base type for all connections and derives
the types for CAN and IP connections from it, allowing function
pointers to be used to simplify the main function.