Use a in6_connection structure in client mode
When running in client mode, canny does not use the same structures
as in server mode, but instead handles buffers and file descriptors
individually. To allow the code to be simplified, canny should use
the same data structures regardless of how it is executed.

This commit modifies canny so it uses an in6_connection structure
to handle file descriptors and receive buffers when running as a
client.