Introduce in6_server structure to handle server sockets
When acting as a server, canny does not manage the listening socket
using a connection structure like the other sockets, but rather using
a lose file descriptor.

This commit introduces a new in6_server type and uses it to manage
server connections, allowing the code to be simplified in the future.