Don't refer to connections by their fd when acting as client
When canny acts as a client, it uses lose file descriptors to read
from network connections. Instead, it should add the connection
structure to the epoll set and use whatever is in the structure
that is returned by epoll.

This commit modifies canny so that it does not use lose file
descriptors to handle the server connection when acting as client.