Merge server-mode and client-mode event loop
Canny is implemented in one huge function with two event loops (one
for server mode, one for client mode) and a lot of code duplication
in either event loop. Code duplication can be reduced by merging the
two loops and moving the code within into functions.

This commit moves the code from the event loops into functions and
calls them through pointers in the connection objects, allowing the
event loops to be significantly simplified and merged.