Correctly assign socket when opening a CAN interface
When opening a CAN interface, canny assigns interfaces an incorrect
file descriptor, rendering it unable to receive messages. This is
due to incorrect parentheses in the statement that acquires a CAN
socket and checks its value, causing the result of the check to be
used as file descriptor.

This commit fixes the parentheses, causing the check to occur after
the file descriptor has been assigned.