Store interface name in can_iface structure
The interface name is necessary to query the MTU of an interface
using the SIOCGIFMTU ioctl, but the name of the interface is not
stored in the can_iface structure.

This commit modifies canny so that the interface name is stored in
the can_iface structure of an interface for easy access.