Use naive canny message format to distinguish between frame types
When canny receives raw CAN frames over a network, it has no way to
distinguish between a CAN2.0B frame and a short CANFD frame, which
might cause it to interpret the end of a CANFD frame as the start of
a new CAN frame.
The simplest way to handle this reliably is to use a fixed-length
message format for messages sent over non-CAN sockets.

This commit modifies canny so that it encapsulates CAN frames in a
fixed-length message when sending it over a socket, allowing it to
reliably identify the CAN frame on the receiving side.